requirements.txt
Flask==2.1.0
gunicorn==20.1.0
transformers
torch
librosa
setuptools-rust
openai-whisper
ffmpeg-python
google-api-python-client
grpcio
firebase-admin
protofbuf3
